+++ tests/testthat/test-mutate.r
@@ -725,23 +725,12 @@ test_that("mutate() supports unquoted values", {
 
 # Error messages ----------------------------------------------------------
 
-test_that("mutate fails gracefully on non-vector columns (#1803)", {
+test_that("mutate handles raw vectors in columns (#1803)", {
   df <- data_frame(a = 1:3, b = as.raw(1:3))
-  expect_error(
-    mutate(df, a = 1),
-    "Column `b` is of unsupported type raw vector",
-    fixed = TRUE
-  )
-  expect_error(
-    mutate(df, b = 1),
-    "Column `b` is of unsupported type raw vector",
-    fixed = TRUE
-  )
-  expect_error(
-    mutate(df, c = 1),
-    "Column `b` is of unsupported type raw vector",
-    fixed = TRUE
-  )
+  expect_identical( mutate(df, a = 1), data_frame(a = 1, b = as.raw(1:3)) )
+  expect_identical( mutate(df, b = 1), data_frame(a = 1:3, b = 1) )
+  expect_identical( mutate(df, c = 1), data_frame(a = 1:3, b = as.raw(1:3), c = 1) )
+  expect_identical( mutate(df, c = as.raw(a)), data_frame(a = 1:3, b = as.raw(1:3), c = as.raw(1:3)) )
 })
 
 test_that("grouped mutate errors on incompatible column type (#1641)", {